+++ database/migrations/2018_02_02_000006_create_products_table.php
@@ -27,8 +27,8 @@ class CreateProductsTable extends Migration
             $table->integer('rate_2')->unsigned();
             $table->integer('rate_3')->unsigned();
             $table->string('address');
-            $table->float('lat');
-            $table->float('lng');
+            $table->decimal('lat', 9, 6);
+            $table->decimal('lng', 9, 6);
             $table->string('image');
 
             $table->timestamps();
